cvi_mpi: weekly rls 2025.05.06
c72c37, [feat](jpege) use input buffer as output buffer when bufferfull, user level
d59646, fix v420 daily build error.
e2bb1a, [feat](isp-af):add regsiter motor cb func sample code

+++ modules/isp/cv181x/isp-tool-daemon/CviIspTool.sh
@@ -88,8 +88,12 @@ elif [ -d "/sys/module/cv180x_vcodec/" ];then
 echo 0x20001 > /sys/module/cv180x_vcodec/parameters/vcodec_mask
 fi
 # enable remap
+if [ -f "/sys/module/cvi_vc_driver/parameters/addrRemapEn" ]; then
 echo 1 > /sys/module/cvi_vc_driver/parameters/addrRemapEn
+fi
+if [ -f "/sys/module/cvi_vc_driver/parameters/ARExtraLine" ]; then
 echo 256 > /sys/module/cvi_vc_driver/parameters/ARExtraLine
+fi
 
 if [ "$HOST" ]; then
     if [ "$GIGABIT" == "true" ]; then
